AlertFactory and AlertFactoryType do not depend on viewController anymore
If you want to create an alert from zero, you may need to specify the alertController: UIViewController when you associated the AlertFactoryType with your class. This is necessary because AlertFactory needs the viewController to present on window. If you are extending some viewController with AlertFactoryType, you don't need to implement alertController getter.